Add Sauce Labs configurations for testing on Safari, IE, Edge and Mobile
Add Sauce Labs remote WebDriver configurations for
testing on browsers that are not available on Linux.

The Sauce Labs configurations will require the local web
server to be exposed via a suitable tunnel (eg. Sauce Connect
or alternatively something like ngrok).

The IE and Edge configurations currently fail
because the 'loggingPrefs' Selenium configuration option
is not supported.

The mobile configurations are also currently failing with
errors to be investigated.